+++ android/app/build.gradle
@@ -30,6 +30,8 @@ android {
         versionName = flutter.versionName
     }
 
+
+
     buildTypes {
         release {
             // TODO: Add your own signing config for the release build.
